Feat: Add model import button to Vue model list popover.
