+++ src/platform/assets/components/AssetBrowserModal.vue
@@ -35,7 +35,7 @@
           size="md"
           class="!h-10 [&>span]:hidden md:[&>span]:inline"
           :label="$t('assetBrowser.uploadModel')"
-          :on-click="handleUploadClick"
+          :on-click="showUploadDialog"
         >
           <template #icon>
             <i class="icon-[lucide--package-plus]" />
@@ -70,17 +70,14 @@ import IconTextButton from '@/components/button/IconTextButton.vue'
 import SearchBox from '@/components/input/SearchBox.vue'
 import BaseModalLayout from '@/components/widget/layout/BaseModalLayout.vue'
 import LeftSidePanel from '@/components/widget/panel/LeftSidePanel.vue'
-import { useFeatureFlags } from '@/composables/useFeatureFlags'
 import AssetFilterBar from '@/platform/assets/components/AssetFilterBar.vue'
 import AssetGrid from '@/platform/assets/components/AssetGrid.vue'
-import UploadModelDialog from '@/platform/assets/components/UploadModelDialog.vue'
-import UploadModelDialogHeader from '@/platform/assets/components/UploadModelDialogHeader.vue'
 import type { AssetDisplayItem } from '@/platform/assets/composables/useAssetBrowser'
 import { useAssetBrowser } from '@/platform/assets/composables/useAssetBrowser'
+import { useModelUpload } from '@/platform/assets/composables/useModelUpload'
 import type { AssetItem } from '@/platform/assets/schemas/assetSchema'
 import { assetService } from '@/platform/assets/services/assetService'
 import { formatCategoryLabel } from '@/platform/assets/utils/categoryLabel'
-import { useDialogStore } from '@/stores/dialogStore'
 import { useModelToNodeStore } from '@/stores/modelToNodeStore'
 import { OnCloseKey } from '@/types/widgetTypes'
 
@@ -95,7 +92,6 @@ const props = defineProps<{
 }>()
 
 const { t } = useI18n()
-const dialogStore = useDialogStore()
 
 const emit = defineEmits<{
   'asset-select': [asset: AssetDisplayItem]
@@ -189,25 +185,5 @@ function handleAssetSelectAndEmit(asset: AssetDisplayItem) {
   props.onSelect?.(asset)
 }
 
-const { flags } = useFeatureFlags()
-const isUploadButtonEnabled = computed(() => flags.modelUploadButtonEnabled)
-
-function handleUploadClick() {
-  dialogStore.showDialog({
-    key: 'upload-model',
-    headerComponent: UploadModelDialogHeader,
-    component: UploadModelDialog,
-    props: {
-      onUploadSuccess: async () => {
-        await execute()
-      }
-    },
-    dialogComponentProps: {
-      pt: {
-        header: 'py-0! pl-0!',
-        content: 'p-0!'
-      }
-    }
-  })
-}
+const { isUploadButtonEnabled, showUploadDialog } = useModelUpload(execute)
 </script>
